feat(sofi): signup landing and success page
