+++ packages/blockchain-wallet-v4-frontend/src/components/Terms/index.tsx
@@ -4,9 +4,9 @@ import styled from 'styled-components'
 
 import { Link, Text } from 'blockchain-info-components'
 
-const TermsContainer = styled.div<{ isCentered?: boolean }>`
+const TermsContainer = styled.div<{ inline?: boolean; isCentered?: boolean }>`
   & > * {
-    display: inline-block;
+    display: ${(props) => (props.inline ? 'inline' : 'inline-block')};
   }
   ${(props) =>
     props.isCentered &&
@@ -15,7 +15,7 @@ const TermsContainer = styled.div<{ isCentered?: boolean }>`
   `};
 `
 const Terms = (props) => {
-  const { company, isCentered = null, recovery } = props
+  const { company, isCentered = null, recovery, sofi } = props
   switch (company) {
     case 'blockchain-kyc':
       return (
@@ -51,6 +51,44 @@ const Terms = (props) => {
               defaultMessage='Privacy Policy.'
             />
           </Link>
+        </TermsContainer>
+      )
+    case 'sofi':
+      return (
+        <TermsContainer style={{ paddingLeft: '4px' }} isCentered={isCentered} inline>
+          <Text color='grey800' size='12px' weight={500}>
+            <FormattedMessage
+              id='scenes.register.registerform.blockchain.read_1'
+              defaultMessage='By checking this box, I acknowledge that I have read and accept the Blockchain.com'
+            />
+          </Text>
+          <Link
+            href='https://www.blockchain.com/legal/terms'
+            target='_blank'
+            size='12px'
+            weight={500}
+            data-e2e='blockchainTermsLink'
+          >
+            <FormattedMessage
+              id='scenes.register.registerform.blockchain.terms'
+              defaultMessage='Terms of Service'
+            />
+          </Link>
+          &nbsp;
+          <Text color='grey800' size='12px' weight={500}>
+            &
+          </Text>
+          &nbsp;
+          <Link
+            href='https://www.blockchain.com/legal/privacy'
+            target='_blank'
+            size='12px'
+            weight={500}
+            data-e2e='blockchainPrivacyLink'
+          >
+            <FormattedMessage id='copy.privacy_policy' defaultMessage='Privacy Policy' />
+          </Link>
+          .
         </TermsContainer>
       )
     default:
