+++ packages/blockchain-wallet-v4-frontend/src/data/signup/sagas.ts
@@ -383,8 +383,11 @@ export default ({ api, coreSagas, networks }) => {
     const product = queryParams.get('product') as ProductAuthOptions
     const platform = queryParams.get('platform') as PlatformTypes
     const signupRedirect = queryParams.get('redirect') as SignupRedirectTypes
+    const pathname = yield select(selectors.router.getPathname)
+    const isSofi = pathname.includes('sofi')
     yield put(
       actions.signup.setProductSignupMetadata({
+        isSofi,
         platform,
         product,
         referrerUsername,
