+++ packages/blockchain-wallet-v4-frontend/src/scenes/Signup/components/SofiSignupCard/index.tsx
@@ -1,333 +1,77 @@
-import React, { useEffect } from 'react'
+import React, { useState } from 'react'
 import { FormattedMessage } from 'react-intl'
-import { Field, InjectedFormProps } from 'redux-form'
+import { text } from 'express'
+import { InjectedFormProps } from 'redux-form'
 import styled from 'styled-components'
 
-import { CountryScope } from '@core/types'
-import { Button, HeartbeatLoader, Text, TextGroup } from 'blockchain-info-components'
-import CheckBox from 'components/Form/CheckBox'
-import Form from 'components/Form/Form'
-import FormError from 'components/Form/FormError'
-import FormGroup from 'components/Form/FormGroup'
-import FormItem from 'components/Form/FormItem'
-import FormLabel from 'components/Form/FormLabel'
-import PasswordBox from 'components/Form/PasswordBox'
-import SelectBox from 'components/Form/SelectBox'
-import TextBox from 'components/Form/TextBox'
-import Terms from 'components/Terms'
-import { CountryType, SignUpGoalDataType, StateType } from 'data/types'
-import { useCountryList, useUSStateList } from 'hooks'
-import {
-  required,
-  stringContainsLowercaseLetter,
-  stringContainsNumber,
-  stringContainsSpecialChar,
-  stringContainsUppercaseLetter,
-  stringLengthBetween,
-  validEmail,
-  validPasswordConfirmation,
-  validStrongPassword
-} from 'services/forms'
-import { applyToUpperCase } from 'services/forms/normalizers'
+import { Badge, Button, HeartbeatLoader, Text, TextGroup } from 'blockchain-info-components'
+import { PlatformTypes } from 'data/types'
+import { isMobile, media } from 'services/styles'
 
-import { SIGNUP_FORM } from '../..'
 import { SubviewProps } from '../../types'
-import ContinueOnPhone from '../SignupForm/ContinueOnPhone'
+import { Card, CardInfo, CardTitle, CardWrapper, InfoItem, InfoTitle, PaddingWrapper } from '..'
+import QRsModal, { QRModalType } from '../SignupForm/QRsModal'
+import SofiSignupForm from '../SofiSignupForm'
 
-const StyledForm = styled(Form)`
-  margin-top: 20px;
-
-  > div * {
-    max-height: 26rem;
-    transition: all 0.5s ease;
+const Bottom = styled.div`
+  margin: 2rem 0 0;
+  > a {
+    margin: 0 8px;
   }
 `
-const FieldWrapper = styled.div`
-  margin-top: 0.25rem;
-  margin-right: 0 !important;
-`
-const FieldWithoutBottomRadius = styled(FormItem)<{ setBorder: boolean }>`
-  .bc__control {
-    border-radius: ${(props) => (props.setBorder ? '8px 8px 0 0 ' : '8px')};
-  }
 
-  .bc__menu {
-    overflow: hidden;
-  }
-`
-const FieldWithoutTopRadius = styled(FormItem)<{ setBorder: boolean }>`
-  .bc__control {
-    border-radius: ${(props) => (props.setBorder ? '0 0 8px 8px' : '8px')};
-  }
-  .bc__menu {
-    overflow: hidden;
-  }
-`
-const PasswordRequirementText = styled(Text)<{ isValid?: boolean }>`
-  font-size: 12px;
-  font-weight: 400;
-  color: ${(props) => (props.isValid ? props.theme.grey800 : props.theme.red600)};
+const LinkAccountSpacer = styled.div`
+  height: 1rem;
 `
 
-const validatePasswordConfirmation = validPasswordConfirmation('password')
-
-const getCountryElements = (countries: Array<CountryType>) => [
-  {
-    group: '',
-    items: countries.map((country: CountryType) => ({
-      text: country.name,
-      value: country.code
-    }))
-  }
-]
-
-const getStateElements = (states: Array<StateType>) => [
-  {
-    group: '',
-    items: states.map((state: StateType) => ({
-      text: state.name,
-      value: state.code
-    }))
-  }
-]
-
-const SignupForm = (props: Props) => {
+const SofiSignupCard = (props: InjectedFormProps<{}> & SubviewProps) => {
   const {
-    formActions,
-    formValues,
-    goals,
-    initialValues,
-    invalid,
     isFormSubmitting,
-    isReferralEnabled,
-    isValidReferralCode,
-    onCountrySelect,
     onSignupSubmit,
-    setShowModal,
-    showState
+    showForm,
+
+    toggleSignupFormVisibility
   } = props
+  const buttonSubmit = showForm ? onSignupSubmit : toggleSignupFormVisibility
+  const showOnlySignup = showForm
 
-  const passwordValue = formValues?.password || ''
-  const referralValue = formValues?.referral || ''
-  const isStateTexas = formValues?.state === 'US-TX'
-
-  const showReferralError =
-    referralValue.length > 0 && isValidReferralCode !== undefined && !isValidReferralCode
-
-  const { data: supportedCountries } = useCountryList({ scope: CountryScope.SIGNUP })
-  const { data: supportedUSStates } = useUSStateList()
-  const dataGoal = goals.find((g) => g.name === 'signup')
-  const { email }: SignUpGoalDataType = dataGoal?.data || {}
-
-  useEffect(() => {
-    if (email) {
-      formActions.change(SIGNUP_FORM, 'email', email)
-    }
-  }, [formActions, email])
-
-  if (!supportedCountries?.countries || !supportedUSStates?.states) {
-    return <></>
-  }
+  const [showModal, setShowModal] = useState<QRModalType | null>(null)
+  const closeModal = () => setShowModal(null)
 
   return (
-    <StyledForm override onSubmit={onSignupSubmit} initialValues={initialValues}>
-      <FormGroup>
-        <FormItem>
-          <FormLabel htmlFor='email'>
-            <FormattedMessage
-              id='scenes.security.email.verifiedtitle'
-              defaultMessage='Email Address'
-            />
-          </FormLabel>
-          <Field
-            autoFocus
-            bgColor='grey000'
-            component={TextBox}
-            data-e2e='signupEmail'
-            name='email'
-            noLastPass
-            placeholder='Enter Email Address'
-            validate={[required, validEmail]}
-          />
-        </FormItem>
-      </FormGroup>
-      <FormGroup>
-        <FormItem>
-          <FormLabel htmlFor='password' id='password'>
-            <FormattedMessage defaultMessage='Password' id='scenes.register.password' />
-          </FormLabel>
-          <Field
-            bgColor='grey000'
-            component={PasswordBox}
-            data-e2e='signupPassword'
-            name='password'
-            placeholder='Enter Password'
-            validate={[required, validStrongPassword]}
-          />
-        </FormItem>
-        {passwordValue.length > 0 && !!validStrongPassword(passwordValue) && (
-          <div style={{ marginTop: '4px' }}>
-            <TextGroup inline>
-              <PasswordRequirementText isValid>
+    <>
+      {showModal && <QRsModal platform={showModal} onClose={closeModal} />}
+      <CardWrapper hideMargin>
+        <Card>
+          <PaddingWrapper>
+            <CardTitle>
+              <Text size='24px' color='textBlack' weight={600} lineHeight='1.5'>
                 <FormattedMessage
-                  id='scenes.register.password.part1'
-                  defaultMessage='Passwords must contain a'
-                />{' '}
-              </PasswordRequirementText>
-              <PasswordRequirementText isValid={stringContainsLowercaseLetter(passwordValue)}>
-                <FormattedMessage
-                  id='scenes.register.password.part2'
-                  defaultMessage='lowercase letter'
+                  id='scenes.signup.sofi.header'
+                  defaultMessage='Let’s get your account setup.'
                 />
-                {', '}
-              </PasswordRequirementText>
-              <PasswordRequirementText isValid={stringContainsUppercaseLetter(passwordValue)}>
+              </Text>
+              <Text
+                color='textBlack'
+                size='16px'
+                weight={500}
+                lineHeight='1.5'
+                style={{ marginTop: '8px', textAlign: 'center' }}
+              >
                 <FormattedMessage
-                  id='scenes.register.password.part3'
-                  defaultMessage='uppercase letter'
+                  id='scenes.signup.sofi.subtitle'
+                  defaultMessage='Create a password for your Blockchain.com account. Next, we’ll migrate your crypto from SoFi.'
                 />
-                {', '}
-              </PasswordRequirementText>
-              <PasswordRequirementText isValid={stringContainsNumber(passwordValue)}>
-                <FormattedMessage id='scenes.register.password.part4' defaultMessage='number' />
-                {', '}
-              </PasswordRequirementText>
-              <PasswordRequirementText isValid={stringContainsSpecialChar(passwordValue)}>
-                <FormattedMessage
-                  id='scenes.register.password.part5'
-                  defaultMessage='special character'
-                />{' '}
-              </PasswordRequirementText>
-              <PasswordRequirementText isValid>
-                <FormattedMessage
-                  id='scenes.register.password.part6'
-                  defaultMessage='and be at least'
-                />
-              </PasswordRequirementText>
-              <PasswordRequirementText isValid={stringLengthBetween(passwordValue, 8, 64)}>
-                <FormattedMessage
-                  id='scenes.register.password.characters'
-                  defaultMessage='8 characters'
-                />{' '}
-              </PasswordRequirementText>
-              <PasswordRequirementText isValid>
-                <FormattedMessage id='copy.long' defaultMessage='long' />.
-              </PasswordRequirementText>
-            </TextGroup>
-          </div>
-        )}
-      </FormGroup>
-      <FormGroup>
-        <FormItem>
-          <FormLabel htmlFor='confirmationPassword' id='confirmationPassword'>
-            <FormattedMessage
-              defaultMessage='Confirm Password'
-              id='scenes.register.confirmpassword'
-            />
-          </FormLabel>
-          <Field
-            bgColor='grey000'
-            component={PasswordBox}
-            data-e2e='signupConfirmPassword'
-            name='confirmationPassword'
-            placeholder='Re-enter Password'
-            validate={[required, validatePasswordConfirmation]}
-          />
-        </FormItem>
-      </FormGroup>
-      <FormGroup>
-        <FieldWithoutBottomRadius setBorder={showState}>
-          <FormLabel htmlFor='country' id='country'>
-            <FormattedMessage
-              defaultMessage='Country of Residence'
-              id='scenes.register.countryofresidence'
-            />
-          </FormLabel>
-          <Field
-            data-e2e='selectCountryDropdown'
-            name='country'
-            validate={required}
-            elements={getCountryElements(supportedCountries.countries)}
-            component={SelectBox}
-            menuPlacement='auto'
-            onChange={onCountrySelect}
-            label='Select Country'
-          />
-        </FieldWithoutBottomRadius>
-        {showState ? (
-          <FieldWithoutTopRadius setBorder={showState}>
-            <Field
-              name='state'
-              elements={getStateElements(supportedUSStates.states)}
-              component={SelectBox}
-              errorBottom
-              validate={[required]}
-              label='Select State'
-            />
-          </FieldWithoutTopRadius>
-        ) : null}
-        {isStateTexas && (
-          <FormItem>
-            <ContinueOnPhone setShowModal={setShowModal} />
-          </FormItem>
-        )}
-      </FormGroup>
-      {isReferralEnabled && (
-        <FormGroup>
-          <FormItem>
-            <FormLabel htmlFor='referral' id='referral'>
-              <FormattedMessage
-                defaultMessage='Have a referral code?'
-                id='scenes.register.referralcode'
-              />
-            </FormLabel>
-            <Field
-              bgColor='grey000'
-              component={TextBox}
-              data-e2e='referral'
-              name='referral'
-              normalize={applyToUpperCase}
-              placeholder='Enter Referral Code'
-            />
-            {showReferralError && (
-              <FormError data-e2e='referralError' style={{ paddingTop: '5px' }}>
-                <FormattedMessage
-                  defaultMessage='Please enter a valid referral code'
-                  id='scenes.register.referralcode.error'
-                />
-              </FormError>
-            )}
-          </FormItem>
-        </FormGroup>
-      )}
-      <FormGroup inline>
-        <FieldWrapper>
-          <Field name='secretPhase' validate={[required]} component={CheckBox} hideErrors />
-        </FieldWrapper>
-        <FormLabel style={{ marginTop: '1px' }}>
-          <Terms />
-        </FormLabel>
-      </FormGroup>
-      <Button
-        data-e2e='signupButton'
-        disabled={isFormSubmitting || invalid || isStateTexas}
-        fullwidth
-        height='48px'
-        nature='primary'
-        type='submit'
-      >
-        {isFormSubmitting ? (
-          <HeartbeatLoader height='20px' width='20px' color='white' />
-        ) : (
-          <Text color='whiteFade900' size='16px' weight={600}>
-            <FormattedMessage id='buttons.continue' defaultMessage='Continue' />
-          </Text>
-        )}
-      </Button>
-    </StyledForm>
+              </Text>
+            </CardTitle>
+
+            <SofiSignupForm {...props} setShowModal={setShowModal} />
+          </PaddingWrapper>
+          <Bottom />
+        </Card>
+      </CardWrapper>
+    </>
   )
 }
 
-type Props = InjectedFormProps<{}> & SubviewProps & { setShowModal?: (e) => void }
-
-export default SignupForm
+export default SofiSignupCard
